% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{prob_ra}
\alias{prob_ra}
\title{prob_ra}
\usage{
prob_ra(p = 0.5, b = NULL, n = NULL)
}
\arguments{
\item{p}{assignment probability of length n}

\item{b}{block indicator of length n}

\item{n}{Number of units}
}
\description{
Multivalued treatment assigned
}
\examples{
prob_ra(n = 10)

}
