% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{systematic_ra}
\alias{systematic_ra}
\title{systematic_ra function assigns units in ordered sequence}
\usage{
systematic_ra(p = 0.5, n = NULL)
}
\arguments{
\item{p}{assignment probability of length n}

\item{n}{Number of units}

\item{m}{Number of units to be assigned to treatment}
}
\description{
Binary treatment assigned
}
\examples{
systematic_ra(n = 10)
systematic_ra(p = c(.3, .4, .9))
}
