% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic.R
\name{prob_ra_internal}
\alias{prob_ra_internal}
\title{prob_ra function for single treatments}
\usage{
prob_ra_internal(p = 0.5, b = NULL, n = NULL, tol = 10)
}
\arguments{
\item{p}{assignment probability of length n}

\item{b}{block indicator, of length n}

\item{n}{Number of units}

\item{tolerance}{parameter for rounding errors}
}
\description{
Binary treatment assigned in blocks b with probability p
}
\examples{
prob_ra_internal(n = 10)

}
